requirements.txt
Flask==3.1.0
openai==1.76.0
pdfplumber==0.11.6
python-dotenv==1.1.0
Werkzeug==3.1.3
